feat: chained background processing to eliminate capture loop idle gaps (#87)
Split process_wav() into two stages: process_wav_stt() (DSP, STT,
detection) chains sequentially across captures, process_wav_exec()
(DOOM, postcard) fires async so the next STT starts immediately
without waiting for DOOM/postcard. Sequential mode unaffected.

Also: sdr_duration changed from int to double for sub-second captures,
plot scripts fixed to assign capture numbers from file paths instead
of timestamps (fixes Cap #2 mislabel on artifact threads), regenerated
v3/v4 EM plots.

+++ pretty-doomed/src/config.cpp
@@ -60,7 +60,7 @@ static void apply_config_entry(const std::string& key, const std::string& value,
     else if (key == "sdr_gain") cfg.sdr_gain = std::stod(value);
     else if (key == "sdr_fm_deviation") cfg.sdr_fm_deviation = std::stod(value);
     else if (key == "sdr_uri") cfg.sdr_uri = value;
-    else if (key == "sdr_duration") cfg.sdr_duration = std::stoi(value);
+    else if (key == "sdr_duration") cfg.sdr_duration = std::stod(value);
     else if (key == "sdr_max_iq_mb") cfg.sdr_max_iq_mb = std::stoi(value);
     else if (key == "sdr_audio_rate") cfg.sdr_audio_rate = std::stoi(value);
     else if (key == "sdr_lpf_cutoff") cfg.sdr_lpf_cutoff = std::stod(value);
